bits
- catch up on a couple of DOS bits
- fix a copy-and-paste error on hack.c function

+++ src/hack.c
@@ -32,7 +32,7 @@ unsigned ui;
 
 anything *
 long_to_any(lng)
-unsigned lng;
+long lng;
 {
     tmp_anything = zeroany;
     tmp_anything.a_long = lng;
